backend/requirements.txt
Django==4.2
django-cors-headers == 3.13.0
djangorestframework==3.14
djangorestframework-simplejwt==5.2
daphne==4.0.0
requests==2.31
psycopg2-binary==2.9
drf-yasg==1.21.7
celery==5.3.6
redis==4.6.0
channels==4.0.0
channels-redis==4.2.0
whitenoise==6.4.0